rename normalizeMarkdownShoutouts.ts
