+++ ui/leafwiki-ui/src/features/preview/MarkdownPreview.tsx
@@ -40,7 +40,7 @@ import { MarkdownLink } from './MarkdownLink'
 import './markdownPreviewCodeTheme.css'
 import MermaidBlock from './MermaidBlock'
 import { normalizeMarkdownListIndentation } from './normalizeMarkdownListIndentation'
-import { normalizeMarkdownBlocks } from './normalizeMarkdownShoutouts'
+import { normalizeMarkdownBlocks } from './normalizeMarkdownBlocks'
 import { rehypeLineNumber } from './rehypeLineNumber'
 import { rehypeWhitelistStyles } from './rehypeWhitelistStyles'
 import { syntaxHighlightLanguages } from './syntaxHighlightLanguages'
